Move sample components out of tests
