+++ test/pipelines/integration/test_fixed_merging_pipeline.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from pprint import pprint
 
 from canals.pipeline import Pipeline
-from test.sample_components import AddFixedValue, Subtract
+from sample_components import AddFixedValue, Subtract
 
 import logging
