+++ test/pipelines/integration/test_looping_pipeline.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from pprint import pprint
 
 from canals.pipeline import Pipeline
-from test.sample_components import Accumulate, AddFixedValue, Threshold, MergeLoop
+from sample_components import Accumulate, AddFixedValue, Threshold, MergeLoop
 
 import logging
