+++ test/pipelines/integration/test_parallel_branches_pipeline.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from pprint import pprint
 
 from canals.pipeline import Pipeline
-from test.sample_components import AddFixedValue, Repeat, Double
+from sample_components import AddFixedValue, Repeat, Double
 
 import logging
