+++ test/sample_components/test_add_value.py
@@ -1,40 +1,8 @@
 # SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
-from typing import Optional
-
-
-from canals import component
-from canals.testing.test_component import BaseTestComponent
-
-
-@component
-class AddFixedValue:
-    """
-    Adds the value of `add` to `value`. If not given, `add` defaults to 1.
-    """
-
-    @component.input  # type: ignore
-    def input(self):
-        class Input:
-            value: int
-            add: int
-
-        return Input
-
-    @component.output  # type: ignore
-    def output(self):
-        class Output:
-            value: int
-
-        return Output
-
-    def __init__(self, add: Optional[int] = 1):
-        if add:
-            self.defaults = {"add": add}
-
-    def run(self, data):
-        return self.output(value=data.value + data.add)
+from canals.testing import BaseTestComponent
+from sample_components import AddFixedValue
 
 
 class TestAddFixedValue(BaseTestComponent):
